modificaciones para visualizacion de desarrollador
